Added picker view to pick the number of Dice.
I MUST add a label for the picker view.

THEN, I must VARY the x and y path on the way to the top and just fix the x path on the way down so it appears they all get jumbled and then break apart and ensure they do not touch.